pyproject.toml
[tool.poetry]
name = "agents"
version = "0.1.0"
description = ""
authors = ["mfmezger <marc.mezger@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
scrapegraphai = "^1.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
